rudimentary edit function for professors, fixed repository find to use right primary key type

+++ src/main/java/edu/hm/cs/fs/scriptinat0r7/repositories/LectureRepository.java
@@ -9,7 +9,7 @@ import edu.hm.cs.fs.scriptinat0r7.model.Lecture;
 /**
  * Repository to execute CRUD operations on {@code Lecture}s.
  */
-public interface LectureRepository extends PagingAndSortingRepository<Lecture, Long> {
+public interface LectureRepository extends PagingAndSortingRepository<Lecture, Integer> {
 
     /**
      * Finds all {@code Lecture}s.
